requirements.txt
requests==2.24.0
pillow==8.0.1
instabot==0.117.0
python-dotenv==0.15.0
urllib3==1.25.10